statarb/requirements.txt
# StatArb - Cointegration 기반 통계적 차익거래
# Python 3.13+ (Anaconda)

# 핵심 의존성
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# 시계열 / 통계 분석
statsmodels>=0.14.0          # cointegration, ADF, Johansen, Markov-switching
hmmlearn>=0.3.0              # Hidden Markov Models
arch>=6.0                    # GARCH 변동성 모델링

# 머신러닝
scikit-learn>=1.3.0          # SVR, GPR, KernelPCA, RandomForest

# DB
psycopg2-binary>=2.9.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.13.0

# 환경 설정
python-dotenv>=1.0.0
